Prevent thread safety issue and improve loading operations to be async
